@@ -17,8 +17,8 @@
     "tauri:prod:build": "yarn tauri:prod build"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "7.0.0",
-    "rollup": "1.29.1",
+    "@rollup/plugin-node-resolve": "7.1.1",
+    "rollup": "1.31.0",
     "rollup-plugin-commonjs": "10.1.0",
     "rollup-plugin-livereload": "1.0.4",
     "rollup-plugin-svelte": "5.1.1",
@@ -26,6 +26,6 @@
     "sirv-cli": "0.4.5"
   },
   "dependencies": {
-    "svelte": "3.17.2"
+    "svelte": "3.18.1"
   }
 }
