@@ -14,6 +14,7 @@ interface DirInfo {
   children?: DirInfo[]
 }
 
+/* eslint-disable security/detect-non-literal-fs-filename */
 function dirTree(filename: string): DirInfo {
   const stats = fs.lstatSync(filename)
   const info: DirInfo = {
@@ -92,6 +93,7 @@ function printAppInfo(tauriDir: string): void {
             // @ts-ignore
             const tauriTomlContents = toml.parse(tauriTomlFile)
             return chalk.green(
+              // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
               `${tauriTomlContents.package.version} (from source)`
             )
           } catch (_) {}
@@ -166,17 +168,19 @@ module.exports = () => {
     key: '  cargo',
     value: getVersion('cargo', [], output => output.split(' ')[1])
   })
-  printInfo({ key: '  tauri-cli', value: getVersion('cargo', ['tauri-cli']) })
+  printInfo({ key: '  tauri-bundler', value: getVersion('cargo', ['tauri-bundler']) })
   printInfo({ key: 'Global packages', section: true })
   printInfo({ key: '  NPM', value: getVersion('npm') })
   printInfo({ key: '  yarn', value: getVersion('yarn') })
   printInfo({ key: 'App directory structure', section: true })
 
   const tree = dirTree(appDir)
-  for (const artifact of tree.children || []) {
+  for (const artifact of tree.children ?? []) {
     if (artifact.type === 'folder') {
       console.log(`/${artifact.name}`)
     }
   }
   printAppInfo(tauriDir)
 }
+
+/* eslint-enable security/detect-non-literal-fs-filename */
