@@ -5,25 +5,26 @@
 
 mod cmd;
 
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
+use serde::Serialize;
 
 use std::io::BufRead;
 
 fn main() {
   tauri::AppBuilder::new()
-    .setup(|_webview| {
-      let handle1 = _webview.handle();
+    .setup(|webview, _| {
+      let handle1 = webview.handle();
       std::thread::spawn(move || {
-        let resource_dir = tauri::api::platform::resource_dir().expect("failed to get resource dir");
+        let resource_dir =
+          tauri::api::platform::resource_dir().expect("failed to get resource dir");
         let node_package_path = resource_dir.join("resources/packaged-node.js");
         let stdout = std::process::Command::new("node")
-          .args(vec!(node_package_path))
+          .args(vec![node_package_path])
           .stdout(std::process::Stdio::piped())
           .spawn()
           .expect("Failed to spawn packaged node")
-          .stdout.expect("Failed to get packaged node stdout");
+          .stdout
+          .expect("Failed to get packaged node stdout");
+
         let reader = std::io::BufReader::new(stdout);
 
         reader
@@ -34,7 +35,7 @@ fn main() {
           });
       });
 
-      let handle2 = _webview.handle();
+      let handle2 = webview.handle();
       tauri::event::listen(String::from("hello"), move |msg| {
         #[derive(Serialize)]
         pub struct Reply {
