@@ -6,7 +6,8 @@ mod runner;
 
 pub struct App {
   invoke_handler: Option<Box<dyn FnMut(&mut WebView<'_, ()>, &str)>>,
-  setup: Option<Box<dyn FnMut(&mut WebView<'_, ()>)>>,
+  setup: Option<Box<dyn FnMut(&mut WebView<'_, ()>, String)>>,
+  splashscreen_html: Option<String>,
 }
 
 impl App {
@@ -23,19 +24,24 @@ impl App {
     }
   }
 
-  pub(crate) fn run_setup(&mut self, webview: &mut WebView<'_, ()>) {
+  pub(crate) fn run_setup(&mut self, webview: &mut WebView<'_, ()>, source: String) {
     match self.setup {
       Some(ref mut setup) => {
-        setup(webview);
+        setup(webview, source);
       }
       None => {}
     }
+  }
+
+  pub fn splashscreen_html(&self) -> Option<&String> {
+    self.splashscreen_html.as_ref()
   }
 }
 
 pub struct AppBuilder {
   invoke_handler: Option<Box<dyn FnMut(&mut WebView<'_, ()>, &str)>>,
-  setup: Option<Box<dyn FnMut(&mut WebView<'_, ()>)>>,
+  setup: Option<Box<dyn FnMut(&mut WebView<'_, ()>, String)>>,
+  splashscreen_html: Option<String>
 }
 
 impl AppBuilder {
@@ -43,6 +49,7 @@ impl AppBuilder {
     Self {
       invoke_handler: None,
       setup: None,
+      splashscreen_html: None,
     }
   }
 
@@ -54,8 +61,13 @@ impl AppBuilder {
     self
   }
 
-  pub fn setup<F: FnMut(&mut WebView<'_, ()>) + 'static>(mut self, setup: F) -> Self {
+  pub fn setup<F: FnMut(&mut WebView<'_, ()>, String) + 'static>(mut self, setup: F) -> Self {
     self.setup = Some(Box::new(setup));
+    self
+  }
+
+  pub fn splashscreen_html(mut self, html: &str) -> Self {
+    self.splashscreen_html = Some(html.to_string());
     self
   }
 
@@ -63,6 +75,7 @@ impl AppBuilder {
     App {
       invoke_handler: self.invoke_handler,
       setup: self.setup,
+      splashscreen_html: self.splashscreen_html,
     }
   }
 }
